move TraceMetadata
